network: Add dedicated network tests module
Currently we keep all the e2e tests in a single module. It is hard to
navigate it and it is a potential risk of variables and definitions
overflowing between modules.

This patch introduces a "network" module under "tests". This module will
contain all the networking e2e tests. This module is "included" into the
main test suite, so it shares preparation, cleanup and artifacts
collection code with the rest of the tests.

The network tests directory starts empty. The reason is that I want to
use this opportunity to refactor and cleanup bloated and chaotic
networking tests. We will move tests from existing modules to new ones
one by one.

Signed-off-by: Petr Horacek <phoracek@redhat.com>