Remove exec readiness probe on the containerDisk container
exec probes cause a lot of problems on various kernel and container
runtime implementations. Getting rid of this exec probe since it is no
longer needed. virt-handler will retry it the containerDisk socket was
not created (which should pretty much never happen).

In the past the exec probe was necessary, since we really had to copy a
disk around. Today the containerDisk binary only needs to create a
socket and sit there.

Signed-off-by: rmohr <rmohr@redhat.com>