selinux: create the tap device using launcher selinux label
This patch reads the selinux labels of both virt-handler and
virt-launcher.

It then proceeds to create the tap-device using the label of
virt-launcher, so it can be consumed by libvirt without requiring
any special selinux policies.

Once that is done, the original context is reset, so all other
binaries can be executed in the virt-handler selinux context.

Calls to SetExecLabel should be wrapped in Lock /Unlock OS thread
until the program execution finishes, to guarantee another
goroutine does not migrate to the current thread before execution
is complete, as per
https://godoc.org/github.com/opencontainers/selinux/go-selinux#SetExecLabel

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>