Support VMI scheme multi IPs list in case of dual stack
In case of dual stack, support listing both IPs under
VMI status.interfaces.ipaddresses.

This commit supports automatic detection of the primary ip
according PodIP, as k8s API enforce that PodIps[0] equal
PodIP which is the primary IP.
The order of the ips and the primary ip in the VMI scheme
will be the same as the ips and primary ip of the pods.

Signed-off-by: Or Shoval <oshoval@redhat.com>