tests, job: Rename RenderJob to NewJob and expose new args
Expose 3 new arguments that control the job behavior, allowing other
users to explicitly control the job:

retry: The number of times the job should try and run the pod.

ttlAfterFinished: The period of time between the job finishing and
its auto-deletion.

timeout: The overall time at which the job is deleted, regardless of it
finishing or not.

Signed-off-by: Edward Haas <edwardh@redhat.com>