tests, job: Convert WaitForJobTo* to a non-assert version.
Waiting for the job to finish is needed not only as an assert helper but
also as a simple blocker function which the caller may not want to
assert at all (the outcome is not of interest).

Converting the current WaitForJob* helpers to a non-assert version and
leaving the assertion for the caller to use if needed.

The new helper introduces additional advantages:
- Collect relevant information through the errors returned (job status).
- Duplicate code is reduced.
- The assertion is now accurately reporting the right line number in the
test and there is no need to use the assertion offset which does not
support well calls from different stack levels.

Signed-off-by: Edward Haas <edwardh@redhat.com>