Fix verifying make targets
Make evaluates conditions in the preprocessing phase - before any target
is executed. It means that we will not be able to detect the changes
made by make or make generate in a make condition.

This change fixes this issue by moving the check into a dedicated script
and calling it from the verifying targets.

We also now sync kubevirt ci in the dedicated container to avoid issues
with different tool versions.

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>