tests, Add waitForNetworkPolicyDeletion
Since the nature of Network Policy logic,
where traffic accepting or denying depends on
all the existing policies,
we should make sure the policy is deleted before
we progress.
Else unexpected flakiness might occur
because the tests assume the only existing policy
is the one that was created by their context block.

Signed-off-by: Or Shoval <oshoval@redhat.com>