selinux: allow creating VMIs on nodes without selinux
We have 2 operations that assume selinux is *not* disabled:
  - read the virt-launcher current selinux context
  - set that context on virt-handler, to create the tap device
    in that scope.

Both of those operations fail when selinux is disabled on a node.

This commit changes the logic on the network handler to read
whether or not selinux is disabled; whenever selinux is disabled,
we do not attempt to perform neither of the aforementioned
operations.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>