Fix sync of generated client-go to master
In the current state, we never update the master branch of
kubevirt/client-go repository. Even we try to push there with `git push
origin master`, the push is never applied and exits with:

+git push origin master
Everything up-to-date

The reason for that is, that our chenges are checked out on branch
"master-local" while the aforemenetioned command tries to push from the
checked out master branch.

With this change, we explicitly push from the currently checked-out
branch to master.

Signed-off-by: Petr Horacek <phoracek@redhat.com>