virt-launcher, Add mechanism to guard add/delete events channel
Since the channel is meant only for the first add and delete
events, its better to protected it, because only the first
events are needed, and the delete event will be consumed
only after the qemu PID is gone.
So this channel should not receive unexpected events,
because it will cause a deadlock.

Signed-off-by: Or Shoval <oshoval@redhat.com>