tests, migration, stress-test: remove doubled `\n`
The safe expecter removes the `\n` from the end of the
previous bSend, since it is translated to `\r\n` and it is
preventing the safe expecter to attach the previous bSend
to it's bExp.

The stress test command has double `\n`, the first one is
removed but the second one is kept, and it is translated to `\r\n`
and the safe expecter fails.

This commit removes the second \n, which is redundant since the tests
passes without it.

Signed-off-by: alonSadan <asadan@redhat.com>