tests, login: expect fedora full prompt
With only the partial prompt (`$`), the expecter can find a match,
in the last case, before it actually got to the login ptompt.
for expamle I encountred  a "#" appearing in the ssh key draw and matched the last case).

This commit replaces the partial prompt with full one to avoid that issue.

Signed-off-by: alonSadan <asadan@redhat.com>