tests, pausing test: increase time for long process
To test pausing, a sleep process is executed on the vmi, then the vmi is paused,
, and the sleep process should still be running after resuming the vmi. To check that
a `sleep 8` process is used, and a `pgrep -f "sleep 8" command should return the `sleep 8` pid.
sometimes when those tests run, the process ends before the call for pgrep,
probably due some delays before and after the pausing phase, and nothing is returned and tests fail.

In this commit the sleeping time is increased to 1 minute which I hope will be sufficient for a reliable test.

Signed-off-by: alonSadan <asadan@redhat.com>