dual stack, tests: ping first on helloWorld{UDP|HTTP} jobs
For whatever reason, ipv6 services get a connection refused when
netcat attempts to connect.

Pinging first, then connecting makes it work.

This is a work-around, which was introduced previously on [0] only
for TCP protocol.

This bug is tracked on [1].

The services tests, in which the exposed services are accessed via
name, had to be updated, receiving the IPFamily over which the test
is supposed to operate.

[0] - https://github.com/kubevirt/kubevirt/commit/d380576d2937b86fcd2cff8a73976fd32186a02e
[1] - https://github.com/kubevirt/kubevirt/issues/4428

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>