virt-handler, sriov: Add network name in VMI interface status
Since network name matching is done by mapping between the MAC address
and the network name, and that for SRIOV interfaces we don't have such
mapping, there is no way to match between the mac which is reported by the
guest agent and the network name, unless the MAC is specified in the spec.
The reason for this limitation is that unlike under interface types,
SRIOV is defined in the libvirt domain using hostdev which doesn't
include mac specification.
The mac is used as a mapper for the network, hence without it
one cannot map between the network name and the reported interface.

Add logic that in case the MAC exists in the spec of the SRIOV,
it would be used to match the right network name
to the interface reported by the guest agent, and appear in
the VMI status.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1856347

Signed-off-by: Or Shoval <oshoval@redhat.com>