tests, console: Replace some NewExpecter usages
Replace the usages of NewExpecter in tests where possible.

It is not possible to replace the helper when the callers are using:
- The returned channel.
- The expecter (in order to examine the batch result).

The incentive here is to reduce the direct usage if this helper in order
to simplify the caller code so it will not need to handle the
teardown/close of the expecter.

Signed-off-by: Edward Haas <edwardh@redhat.com>