macvtap, migration, tests: add a test w/ traffic
Test the VM has connectivity before and after VM migration using
IPv4 addresses. IPv6 is out of scope, but it is also expected to
work out of the box.

We rely on guest agent to learn the IP addresses of the
instantiated VMs; as such, we needed to use a fedora VM for the
migration test with traffic.

The CirrOS VMIs used throughout the tests *only* have a single
network interface, and as such, are not connected to the pod
network.

The Fedora VMI - which is used only in the migration test with
traffic - on the other hand, has 2 network interfaces:
  - a masquerade iface, to connect the VMI to the k8s services
  - a macvtap iface, to connect the VMI to the host's subnet

The masquerade interface, on the Fedora VMI (pod interface) is
required to reach the `cdi-http-importer` k8s service, from where
we install the guest-agent, which, in turn, is required to provide
the test framework the VMI IP.

We identify the macvtap interface based IP based on the mac
address, which we statically provisioned.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>