Remove authorized_key file merging.
KubeVirt now completely controles the authorized_keys file when dynamic
propagation is used. We can consider in the future to expand this
functionality to allow merging. However, for now the complexity of
merging outweighs the benefit. The simplest and most maintainable approach
is to completely sync the authorized keys file from the secret source

Signed-off-by: David Vossel <davidvossel@gmail.com>