tests, multus_tests: make helpers return an error
It is considered best practice to not assert in helper function,
but to return an error instead.
One reason for that is the ability to act upon
received error, like in a negative test for example.
In the multus-test case we might want in the future to remove a NIC,
and check it was removed, and we can use the returned
error from `checkInterface` to assert that.

For that, In this commit, some of the helpers were refactored
to return an error.

Also this commit  exported the `ip link set <interface> up` to a function
of It's own to avoid the repeated code, and adapted the new general prompt expression.

Signed-off-by: alonSadan <asadan@redhat.com>