tests, multus-tests, SRIOV: configure IP based on MAC or name
In the multus tests suit, when configuring ip to an interface,
the name of the interface given to the `ip addr add` command is
hard-coded.

This commit adds a function to configure SRIOV interface
based on the MAC of the interface ,rather than just on It's name.

This allows a test writer to "mark" an SRIOV interface with  a static mac,
and by that to avoid assigning IP to the wrong non-SRIOV
interface. This scenario probably what causes connectivity issues
in the ping tests over SRIOV, causing it to be flaky.

Signed-off-by: alonSadan <asadan@redhat.com>