Fix double migration during node drain
This bug originates from a race condition between the time we finish a
migration and remove the PDB that used to protect the original launcher
and the eviction requests that are issued by the client every 5 seconds.

To fix the issue, a check was added in the eviction webhook that
prevents it from operating on a VMI that was already migrated to another
node.

This PR also includes closing of loose ends on the evacuation
controller to prevent a race between the time we mark the VMI's
migration as completed and the time the evacuation sync loop runs.

https://bugzilla.redhat.com/show_bug.cgi?id=1888790

Signed-off-by: Daniel Belenky <dbelenky@redhat.com>