virt-launcher: introduce dhcpv6
The server will send only the ip address information.

searchDomains, MTU and customDHCPOptions will be handled by the
dhcpv4 server.

Notes-
- We assume that the ipv4 dns server will serve ipv6 as well.
- Routes are relevant to bridge binding only, since curretly the dhcp
server supports only masquerade, routes are not passed.
- This information is taken from the pod and passed to the dhcpv4 server as
is.
It is not validated to fit ipv6.
The assumption is that the validation was done by the CNI.
(i.e if we don't validate the MTU fits ipv6).


Signed-off-by: Alona Kaplan <alkaplan@redhat.com>