tests, utils: shorten name of random VMs
I find it very hard to sift through CI logs with VM names that include 48 pseudorandom consonants. Moreover, kubevirt prepends "virt-launcher-" to the 55-chars-long VMI name and uses this a generateName for the virt-launcher Pod. Apparently, when Kubernetes generates the Pod name, it keeps only 58 chars of the generateName, which makes matching virt-launcher pods even harder.

This commit shortens the VMI name length to a more humanly comprehensible length of 12 consonants. The probability of collision is still quite low (27^12 ~ 10^17). If we keep creating a VMI per second, we are likely to see a collision after a few billion years. The number 12 was chosen without a very good reason. 6 consonants would be safe enough. A dash is introduced between the testvmi prefix and the random string for readability.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>