tests, libvmi, vmi: shorten random vm names #2
As mentioned in the review for PR #4605, we have few other places where
we generate unusably long VM names. This commit makes RandName behave
more like Pod's generateName and add only 5 random consonants.
Hopefully, all other places where we generate random VM names can be
fixed to use this function.

This commit should not be taken as an advocacy for using randomness in
tests. It would be better to have tests with constant meaningful objects
names.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>