virt-launcher, converter: Remove vCPU dependency on queue limits
The number of vCPU is calculated from the CPU topology and the number of
queues for a vNIC is based on the vCPU count.

The current implementation limits the number of vCPU/s and
queues based on the vNIC limitation (tap device maximum queues).

While the queue count logic is correct (dependent on the vCPU count and
the tap device limit), the vCPU count is wrongly limited by the queues
limit.
This change drops the vCPU count limit, keeping it dependent only on the CPU
topology.

Signed-off-by: Edward Haas <edwardh@redhat.com>