masquerade, migration: hardcode the bridge MAC addr
Since the bridge MAC address no longer changes, the ARP tables do
not need to stabilize, which preserves connectivity from the
outside world to the guest.

This is tested in this commit since the `Eventually` block in the
vmi_networking_test for masquerade binding migration can be safely
removed.

The lowest locally administrated unicast MAC address was chosen for
the bridge.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>