tests, sriov: XFail IPv4 connectivity test
The SRIOV connectivity check between SRIOV vnic/s is failing from time
to time [1].
This is happening now for IPv4 in addition to the existing IPv6 test
which is already marked as XFail.

This test is set to XFAIL, i.e. it is left running and in case of
failure, a skip is raised (instead of a failure).
This enables further investigation of the problem.

1. https://github.com/kubevirt/kubevirt/issues/4642

Signed-off-by: Edward Haas <edwardh@redhat.com>