virt-launcher, sriov: Add an alias to the domain device
Once a host-device is added to the domain spec, it is challenging to
identify if it is an SRIOV device and to which VMI interface/network it
belongs to.

In order to allow a simple identification of such devices when reading
the spec from libvirt, the alias field is used to mark a device.

The alias values is added with the following format:
"sriov-<network-name>"

Signed-off-by: Edward Haas <edwardh@redhat.com>