infocache: place code in a new file
This commit moves writeToCachedFile and readFromCachedFile into
a new file, cacheinfo, in the network package. It adds dedicated wrapper
functions for reading/writing virt-launcher and virt-handler "cache
files", and adds few tests for them.

I believe that naming the functions properly, giving them a more
specific signature and moving them to a separate file would open the
path for further refactoring. In the future, I hope that something other
than files can be used for virt-handler to virt-launcher communication.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>