ephemeral-disk-utils: simplify and rename RemoveFile
We would like RemoveFile to remove a file that may already exists from a
former attemt to reconcile a resource. We do not want it to do anything
more. E.g. if it finds a whole directory tree there, it is probably a
programmatic error that should be reported as such.

This commit renames the function to RemoveFilesIfExist, to make the
intended functionality clear, and modify the implemenetation to be more
careful and return an error if the on-disk situation is unexpected.

It also extends the signature to allow this type of optional removal of
multiple files.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>