virt-launcher, sriov: Simplify PCI address decoding from env
The value of the resource in the environment contains a list of PCI
addresses that are associated to it. The list contains PCI addresses
separated by comas.
Unfortunately it may appear with a tailing coma on the last address.

This change simplifies the parsing code that is required to ignore that
trailing coma.

Signed-off-by: Edward Haas <edwardh@redhat.com>