automation, Use cluster-sync instead its partial targets
Use make cluster-sync as this is the official target.
Remove the until, in case the cluster-sync is failed,
we shouldn't mask it with an infinite retry,
nor try forever in case there is a non transient failure,
consuming the resources of the CI.

The reasons it was used are outdated, as we don't have
a OKD provider anymore under kubevirtci.

Signed-off-by: Or Shoval <oshoval@redhat.com>