Don't retry jobs that are expected to fail
At network service testing the timeout for jobs failing or not is 90
seconds and those jobs have three retries, in case the test is expected
to fail 4 differenct containers are going to be create/run/delete since
the job is going to retry it and this reaches the timeout limit (more
often at k8s >= 1.20).

Since on test that expect job to fail retries does not make sense this
change add a parameter to the job creatin function so we specify no
retries if job is expected to fail.

Signed-off-by: Quique Llorente <ellorent@redhat.com>