virt-handler: Block migration for SRIOV VM's
Currently the code for handling Live Migration for SRIOV VM's
on the source and target nodes in not merged yet.
In order to avoid confusion for Kubevirt users and contributors,
Live Migration for VM's with SRIOV interfaces should be blocked
regardless of the feature-gate state.

Once all the bits for SRIOV Live Migration are merged
it should be enabled, and determined by the feature-gate state.

Signed-off-by: Or Mergi <ormergi@redhat.com>