network, bind mech: load/set cached vif uses struct attrs
The PID must still be passed as an argument, to allow unit tests,
since the production code caches the VIF in one path, and reads the
cache from another (symlinked) path. On unit tests, this symlink
does not exist, and is overkill.

The new cache interface mechanism should be extended in a follow
up PR to also cache the VIF structures passed between processes.

In this PR, we also simplify the return types of the methods; it
now replies only with an error - the `exists` boolean it returned
made no sense.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>