Add ephemeral storage request for compute and volumedisk containers
This commit fixes the following bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1731819

If ephemeral storage request is not set in manifest,
it will be automatically set to "1Mi" default for every container.

Also add unit test to verify ephemeral storage request is set. In
addition, modify "..verifying that resources are set.." unit test
so it tests by checking for specific resources instead of length
(test failed since another resource is requested).

Signed-off-by: Itamar Holder <iholder@redhat.com>