tests, Enable watching for vmi errors in case ignoreWarnings is true
E2e tests use waitForVMI in order to check the vmi is starting.
It has two variants:
1. Log warnings, assert on errors.
2. Assert both on warnings and errors.

In case ignoreWarnings is true, it didn't create a watcher at all,
so even vmi errors were not asserted / logged by the test's watcher.

Fix it by always creating the watcher, and setting watcher's
FailOnWarnings to be enabled, in case ignoreWarnings is false.

Signed-off-by: Or Shoval <oshoval@redhat.com>