virt-launcher post migration: Filter host-devices to attach
On post-migration phase on virt-handler VM controller,
in case a request is re-enqueued finalize migration
command will be invoked again and eventually also AttachDevices.
In case at least one host-device of a VMI already attached,
trying to attach it again will raise errors from libvirt.

In order to prevent such errors this change adds a filtering
to attach SRIOV host-devices logic so that only SRIOV host-devices
that are not already connected will take in account.

Signed-off-by: Or Mergi <ormergi@redhat.com>