virt-launcher, cli: Add volatile callback de/register methods
Libvirt connection wrapper exposes several register methods and no
deregister one. The callback registers, except of registering a callback
with libvirt, also re-connect the connection (if needed) and adds the
callbacks to a pool in order to be able and re-register them in case of
a re-connect in the future.

For some operations (e.g. the host-device detach before a migration)
these additional operations are not needed. A simple volatile
register and a deregister is required, which is free of any
side-effects.

This change exposes the VolatileDomainEventDeviceRemovedRegister and
DomainEventDeregister operations.

Signed-off-by: Edward Haas <edwardh@redhat.com>