test, sriov: Explicitly set vlan=0 to NADs with no VLAN defined
There is a difference between a network
(defined in networkAttachmentDefenition)with vlan=0 attribute,
and a network with no vlan attribute at all.
If the VF has some vlan tag different than 0,
then when it is assigned to a pod (or a vmi)
which is connected to a network with vlan=0,
the vlan tag will be removed.
When the pod is connected to a network with no vlan attribute,
the vlan tag will remain as it is.

This behavior is implemented in SRIOV-CNI.
This is where the vlan is assigned :
https://github.com/k8snetworkplumbingwg/sriov-cni/blob/b18123d809f1c010cae19018e3395ed5981e76f7/pkg/sriov/sriov.go#L287

The investigation found that the source of the ping flakiness
is that in our tests, for some reason,
the vlan tag remains from the ping-over-sriov+vlan test.
Then on the ping-over-sriov-no-vlan test,
because the network does not have a vlan defined attribute,
it will not remove the existing vlan tag upon VF assignment.
Then, if unfortunately one VMI will be assigned with a tagged VF,
it will not remove the tag,
and if another VMI assigned with
a VF with no vlan tag ,it will lead to a ping failure.

Another interesting thing
is that when the net-attach-def is created by the SRIOV-operator,
after creating a sriovNetwork object,
it will give the net-attach-def a value of 0,
no matter if the sriovNetwork object has a vlan attribute or not.

By creating the net-attach-def with a vlan=0 attribute,
the vlan tag is removed from the VF the test seems to pass consistently.

Signed-off-by: alonsadan <asadan@redhat.com>