Add support for specifying the logical and physical block sizes of a disk device.
This adds 2 ways to do so:
- Manually setting specific logical and physical sizes
- Detect optimal values based on the backing device

The first is limited to power of 2 values between 512 bytes and 2 MiB (QEMU limits).
The latter is added as a FeatureState to prevent block sizes from suddenly changing/getting added on existing VMs.
If neither is specified, no sizes are set and it defaults to the QEMU defaults as it always has.
Currently only supported for file and block device volumes.

Signed-off-by: Omar Pakker <Omar007@users.noreply.github.com>