tests, outside-conn:add ability to skip and set
The default addresses/domains used in tests for checking connectivity
to the outside world are hard-coded.
When running the tests with no internet connection,
those tests will fail.

It should be possible to run the tests without failures,
also without internet connectivity.
It could be also that the user would like to test outside connectivity
to different addresses/domains then the default.

Depending on the testing method and goal, this commit adds the ability
to skip outside conn tests, or set different addresses and domain to
test against, and adds a description on how to set and skip those tests.

Signed-off-by: alonsadan <asadan@redhat.com>