Simplify greps in determination of new tests
Instead of several greps which will be error prone when renamed files
are in the changeset we just use the --diff-filter and awk print the
last column which always results in the new file name.

Also as test files do not necessarily have to have the _test suffix we
just grep for .go (while still excluding suite files).

Signed-off-by: Daniel Hiller <dhiller@redhat.com>