Add specific nat rules for inbound envoy routing
This commit introduces changes to masquarade binding method that allow
correct Istio proxy sidecar functionality. Only VMI with explicitly
specified list of ports is considered in this commit.

There are two issues preventing istio proxy from working:
1)
dnat rule that forwards traffic to the VM is matched at prerouting,
which makes the inbound traffic to bypass the proxy.
To avoid that, the prerouting dnat rule is not appended, if
Istio is being used, which is decided based on the VMI having
"sidecar.istio.io/inject" annotation set to "true".

2)
Second issue is related to how istio proxy treats traffic targetting
port that is declared in an associated k8s service vs traffic targetting
port which is not declared in associated k8s service.
Traffic that targets port declared in k8s service is sent by the proxy
from 127.0.0.1 to 127.0.0.1:port, which is matched by the existing rules.
However, if the traffic targets port not declared in an associated k8s
service, Istio proxy sends the traffic from 127.0.0.6 to podIP:port.
For these situations, we don't have appropriate rules and so this
traffic is not being forwarded to the VM.

This commit adds additional 2 rules matching traffic that originates from
127.0.0.6 in output and KUBEVIRT_POSTINBOUND chains.

As per discussion https://github.com/kubevirt/kubevirt/pull/5200#discussion_r606840280
there was an attempt to match destination address instead, to avoid
potentionally matching of external traffic, but then we realized, that
this comes with consequence - the KUBEVIRT_POSTINBOUND snat rule would
match too, causing the source address of traffic that reaches VM from
pod network to be 10.0.2.1, which is not desired.

To illustrate the inbound routing process with Istio proxy involved,
see the following diagram depicting proxying packet that targets
port declared in k8s service for the VMI:

__Inbound traffic destined to port 8080 declared in k8s svc__
┌────────────┐
│ k8s network│
└──┬─────────┘
   │
   │
   ▼
 PREROUTING STAGE
┌────────────┐
│ prerouting │(prerouting hook)
└──┬─────────┘
   │
   │ iifname "eth0" counter packets 991 bytes 59460 jump KUBEVIRT_PREINBOUND
   ▼
   ┌──────────────────────┐
   │ KUBEVIRT_PREINBOUND  │  <empty>
   └──────────────────────┘

┌───────────┐
│PREROUTING │ (prerouting hook)
└─┬─────────┘
  │
  │meta l4proto tcp counter packets 150 bytes 9000 jump ISTIO_INBOUND
  ▼
   ┌─────────────┐
   │ISTIO_INBOUND│
   └┬────────────┘
    │
    │meta l4proto tcp counter packets 0 bytes 0 jump ISTIO_IN_REDIRECT
    ▼
   ┌─────────────────┐
   │ISTIO_IN_REDIRECT│ tcp counter packets 0 bytes 0 redirect to :15006
   └───────┬─────────┘
           │
           │
           ▼
       ┌─────────────┐
       │ENVOY PROCESS│ Sends the packet from 127.0.0.1 to 127.0.0.1:8080
       └───┬─────────┘
           │
           │
           ▼
┌─────────┐
│ output  │ ip daddr 127.0.0.1 tcp dport 8080 dnat to 10.0.2.2
└───┬─────┘
    │
    │
    ▼
┌───────────┐
│postrouting│
└───┬───────┘
    │
    │ oifname "k6t-eth0" counter packets 0 bytes 0 jump KUBEVIRT_POSTINBOUND
    ▼
    ┌──────────────────────┐
    │ KUBEVIRT_POSTINBOUND │ tcp dport 8080 ip daddr 10.0.2.2 snat to 10.0.2.1
    └──────────────────────┘

As you can see, in KUBEVIRT_PREINBOUND, we don't dnat, which allows
us to get to ISTIO_IN_REDIRECT, which redirects the packet to port
15006, which is where the proxy listens for inbound traffic.
Proxy sends the packet from localhost to localhost:8080.
(Information about the original dport is retrieved from the socket option
SO_ORIGINAL_DST)

After that, src and dst IP addresses are changed at output and postrouting
chains to forward the proxied packet to the VM.

Another scenario, where the inbound traffic targets port undeclared in a k8s
service (8081 for example), is quite similar. The difference is that proxy sends
the packet from 127.0.0.6 to pod IP address:

     ┌─────────────┐
     │ENVOY PROCESS│ Sends the packet from 127.0.0.6 to podIP:8081
     └─────────────┘

This traffic sent from 127.0.0.6 is matched by the rules added to output
and KUBEVIRT_POSTINBOUND.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>