rpm: Update libvirt and QEMU
Specifically:

  libvirt 6.6.0-13.1 -> 7.0.0-12
     QEMU 5.1.0-18   -> 5.2.0-15

SeaBIOS remains at 1.14.0-1.

Note that, in order to keep aarch64 cross-builds working, a
few additional packages have had to be explicitly added to
the libvirt-devel bazeldnf targets - namely keyutils-libs,
krb5-libs and libmount.

This should not be needed, and I'm not sure what component
would have to be modified so that such hacks are no longer
necessary, but it's consistent with what was done in
https://github.com/kubevirt/kubevirt/pull/5456 when aarch64
cross-builds were introduced.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>