Add unit tests for PodDisruptionBudgets reconciliation
Two unit tests are added in a new file (pdb_test.go).

The tests are for two scenarios. In the first, the cached
PDB and the required PDB are equal. In the second thy are
different. So for the first scenario no patch is expected
and for the second a patch is necessary.

Signed-off-by: Itamar Holder <iholder@redhat.com>