Create test state struct to decouple tests for parallel run
Create a test data struct in order to separate the required test state
(mocks, recording structures etc) from the actual run, so we do not
share state between tests anymore. This eliminates the data races.
Also add the behavior to the struct so that it uses the test state and
not the shared state.
Increase timeouts for tests that regularly fail on parallel execution.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>