probes readiness tests IPv6: Remove Pod/VM not-ready assertion
Currently on dual-stack clusters k8s doesnt enable setting
a readiness probe that will use the pod IPv6.

Due to this constraint on IPv6 readiness probes tests
we create a pod that will act like external service,
It hosts TCP/HTTP server as the readiness prob backend.
Then we create a VMI with a readiness probe with that
pod IPv6 address.

Due to the fact that we create the backend for the prob
before the VMI is up, the assertion for pod/VM not-ready
state is incorrect on IPv6 path.

Signed-off-by: Or Mergi <ormergi@redhat.com>