virt-chroot, tap: Retry the creation of the tap device
The creation of the tap device has been observed failing randomly on CI
with the following error:
`Tuntap IOCTL TUNSETIFF failed [0], errno operation not permitted`

Investigation[1] revealed that a retry is enough to overcome the
failure.

The source of the problem seems to have been handled by BZ#1933627 [2]
and only appears on specific kernel versions.
(between kernel-4.18.0-215.el8 (included) and kernel-4.18.0-297.el8
(not included))

Therefore, a retry mechanism is applied on the tap device creation to
support these kernels (as we do not have control on which kernel
kubevirt is expected to run).
We should consider removing this workaround and add a warning on the
release notes to avoid these version in the future.

This should solve the immediate problem of CI E2E test flakiness.

[1] https://github.com/kubevirt/kubevirt/pull/5390
[2] https://bugzilla.redhat.com/1933627

Signed-off-by: Edward Haas <edwardh@redhat.com>