virt-launcher, Fix flaky VMI report
Virt-handler refresh the VMI domain information,
each 300 seconds by handleResync function.

The information it fetched doesn't include
Guest Agent information such as Guest OS Info,
and interface status.

This will cause flakiness in the VMI status,
each time that the report is fetched,
until the next update of Guest Agent will be sent
by the virt-launcher to the virt-handler.

This commit fix it by decorating the resync
domain info, with the Guest Agent information,
as the virt-launcher itself does before sending
the domain info to the virt-handler.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1907707

Signed-off-by: Or Shoval <oshoval@redhat.com>