refactor, network: Remove podnicfactory
There is a global variale to at virt-launcher/virtwrap/network to mock
the podNIC interface to test the SetupPodPhase1 and SetupPodPhase2
functions, the reality is that only the generation of the nics is tested
and the rest is just passing the domain or the pid from one function to
the other.

This change move podNICImpl to podNIC and move all the code to the
to `getNICs` it also creates a new struct `VMNetworkConfigurator` with
the setup functions.

The only mocked part is the NetworkHandler and it's passed from top
VMNetworkConfigurator to the individual podNIC instances so there is no need for
a factory instance at a global var.

Signed-off-by: Quique Llorente <ellorent@redhat.com>